Moved p5.js functions outside of index.tsx
